home/scripts/python/requirements.txt
python-binance
TA-Lib
numpy~=1.20.2
websocket_client
websocket-client~=0.58.0
